Fixing de-autocast
