+++ nemo/collections/common/parts/training_utils.py
@@ -16,7 +16,7 @@ from contextlib import nullcontext
 
 import torch
 
-__all__ = ['avoid_bfloat16_autocast_context', 'avoid_float16_autocast_context']
+__all__ = ['avoid_bfloat16_autocast_context']
 
 
 def avoid_bfloat16_autocast_context():
@@ -27,23 +27,5 @@ def avoid_bfloat16_autocast_context():
 
     if torch.is_autocast_enabled() and torch.get_autocast_gpu_dtype() == torch.bfloat16:
         return torch.cuda.amp.autocast(dtype=torch.float32)
-    else:
-        return nullcontext()
-
-
-def avoid_float16_autocast_context():
-    """
-    If the current autocast context is float16, cast it to bfloat16
-    if available (unless we're in jit) or float32
-    """
-
-    if torch.is_autocast_enabled() and torch.get_autocast_gpu_dtype() == torch.float16:
-        if torch.jit.is_scripting() or torch.jit.is_tracing():
-            return torch.cuda.amp.autocast(dtype=torch.float32)
-
-        if torch.cuda.is_bf16_supported():
-            return torch.cuda.amp.autocast(dtype=torch.bfloat16)
-        else:
-            return torch.cuda.amp.autocast(dtype=torch.float32)
     else:
         return nullcontext()
